[Fix] 服务端全部更换成 64 位 [Add] 文件操作更换至 File_Opt 类

+++ server/include/tcp.h
@@ -18,10 +18,10 @@ int tcp_init(int port, int max_connection);
 
 int tcp_accept(int epoll_fd, int fd);
 
-int tcp_receive(int client_fd, void *buf, size_t n);
+int tcp_receive(int client_fd, void *buf, int64_t n);
 
-int tcp_send(int sock_fd, char *buffer, size_t length);
+int tcp_send(int sock_fd, char *buffer, int64_t length);
 
-int tcp_sendfile(int sock_fd, int file_fd, off_t *offset, size_t bytes_need_to_send);
+int tcp_sendfile(int sock_fd, int file_fd, off64_t *offset, int64_t bytes_need_to_send);
 
 #endif //FUCK_STORAGE_SERVER_TCP_H
